Operate on image copy rather then on device.
Till now the drill was that image was written on device first, and all
the changes was done directly on the mounted device partitions. This
slow down debugging, so current approach is to have a copy of the image
and on case of device set to destination, it will be a temporary file
deleted after write to device.